fix#63 채팅 및 AI 연동 기능 관련 버그 수정
