+++ src/main/java/org/example/bookmarket/chat/controller/ChatController.java
@@ -8,9 +8,12 @@ import org.example.bookmarket.chat.dto.ChatRoomInfo;
 import org.example.bookmarket.chat.service.ChatService;
 import org.example.bookmarket.common.handler.exception.CustomException;
 import org.example.bookmarket.common.handler.exception.ErrorCode;
-import org.example.bookmarket.security.CustomUserDetails;
+import org.example.bookmarket.user.entity.SocialType;
+import org.example.bookmarket.user.entity.User;
+import org.example.bookmarket.user.repository.UserRepository;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -23,61 +26,91 @@ import java.util.List;
 public class ChatController {
 
     private final ChatService chatService;
+    private final UserRepository userRepository;
 
-    // 📌 채널 생성 또는 조회 (book-detail.html에서 호출)
+    /**
+     * 채널 생성 또는 조회 (book-detail.html에서 호출)
+     */
     @PostMapping("/channel")
     @ResponseStatus(HttpStatus.CREATED)
-    @ResponseBody // JSON 응답 명시
+    @ResponseBody
     public ChatResponse createOrGetChannel(
-            @RequestBody ChatRequest request, // usedBookId, user2Id(판매자ID)가 포함됨
-            @AuthenticationPrincipal CustomUserDetails userDetails // 현재 로그인한 사용자 (구매자)
+            @RequestBody ChatRequest request,
+            Authentication authentication
     ) {
-        // 로그인되지 않은 사용자 처리
-        // [수정] 로그인되지 않은 사용자에 대한 예외 처리를 명확하게 변경합니다.
-        if (userDetails == null || userDetails.getUser() == null) {
-            throw new CustomException(ErrorCode.UNAUTHORIZED_CHAT_ACCESS);
+        User currentUser = resolveCurrentUser(authentication);
+        if (currentUser == null) {
+            throw new CustomException(ErrorCode.LOGIN_REQUIRED);
         }
-        // ChatRequest에 user1Id (구매자 ID)를 설정합니다.
-        // 프론트에서 넘어온 user1Id는 무시하고, 로그인된 사용자 ID를 사용합니다.
-        request.setUser1Id(userDetails.getUser().getId()); // 로그인된 구매자 ID 설정
+        request.setUser1Id(currentUser.getId());
 
-        return chatService.createChannel(request); // createChannel 메서드가 이미 채널 생성/조회 역할 수행
+        return chatService.createChannel(request);
     }
 
-    // 📌 유저의 채널 리스트 조회
+    /**
+     * WebSocket 채팅방 뷰(chatroom.html)를 반환합니다.
+     */
+    @GetMapping("/room/{channelId}")
+    public String getChatRoomPage(@PathVariable Long channelId,
+                                  Model model,
+                                  Authentication authentication) {
+
+        User currentUser = resolveCurrentUser(authentication);
+
+        // [수정] currentUser가 null일 경우를 반드시 확인합니다.
+        if (currentUser == null) {
+            return "redirect:/auth/login";
+        }
+
+        ChatRoomInfo chatRoomInfo = chatService.getChatRoomInfo(channelId, currentUser.getId());
+
+        model.addAttribute("channelId", channelId);
+        model.addAttribute("userId", currentUser.getId());
+        model.addAttribute("partnerNickname", chatRoomInfo.getPartnerNickname());
+        model.addAttribute("bookTitle", chatRoomInfo.getBookTitle());
+        model.addAttribute("bookUrl", "/used-books/" + chatRoomInfo.getBookId());
+
+        return "chatroom";
+    }
+
+    /**
+     * 사용자의 모든 채팅 채널 목록을 조회합니다.
+     */
     @GetMapping("/channels/{userId}")
     @ResponseBody
     public List<ChatResponse> getUserChannels(@PathVariable Long userId) {
         return chatService.getUserChannels(userId);
     }
 
-    // 📌 특정 채널의 메시지 리스트 조회
+    /**
+     * 특정 채널의 모든 메시지 내역을 조회합니다.
+     */
     @GetMapping("/channel/{channelId}/messages")
     @ResponseBody
     public List<ChatMessageResponse> getMessages(@PathVariable Long channelId) {
         return chatService.getMessages(channelId);
     }
 
-    // ✅ chatroom.html 반환 (WebSocket 채팅방 뷰)
-    // URL: /api/chat/room/{channelId}
-    @GetMapping("/room/{channelId}")
-    public String getChatRoomPage(@PathVariable Long channelId,
-                                  Model model,
-                                  @AuthenticationPrincipal CustomUserDetails userDetails) {
-        // 로그인 상태 확인
-        if (userDetails == null || userDetails.getUser() == null) {
-            return "redirect:/login"; // 로그인되어 있지 않다면 로그인 페이지로 리다이렉트
+    /**
+     * 일반 로그인과 소셜 로그인을 모두 처리하여 현재 로그인된 User 엔티티를 반환하는 헬퍼 메서드
+     */
+    private User resolveCurrentUser(Authentication authentication) {
+        if (authentication == null || !authentication.isAuthenticated() || "anonymousUser".equals(authentication.getPrincipal())) {
+            // 로그인하지 않은 경우 null 반환
+            return null;
         }
 
-        // 📌 채널 정보 조회 (상대방 닉네임, 책 제목 등)
-        ChatRoomInfo chatRoomInfo = chatService.getChatRoomInfo(channelId, userDetails.getUser().getId());
+        Object principal = authentication.getPrincipal();
 
-        model.addAttribute("channelId", channelId);
-        model.addAttribute("userId", userDetails.getUser().getId()); // 현재 로그인한 사용자 ID (채팅 메시지 구분용)
-        model.addAttribute("partnerNickname", chatRoomInfo.getPartnerNickname());
-        model.addAttribute("bookTitle", chatRoomInfo.getBookTitle());
-        model.addAttribute("bookUrl", "/books/" + chatRoomInfo.getBookId()); // 책 상세 페이지로 가는 링크
+        if (principal instanceof User) {
+            return (User) principal;
+        } else if (principal instanceof OAuth2User) {
+            OAuth2User oauth2User = (OAuth2User) principal;
+            String socialId = oauth2User.getAttribute("id").toString();
+            return userRepository.findBySocialTypeAndSocialId(SocialType.KAKAO, socialId)
+                    .orElse(null); // 사용자를 찾지 못한 경우 null 반환
+        }
 
-        return "chat/chatroom"; // 👉 templates/chat/chatroom.html 로 연결
+        return null;
     }
 }
