+++ src/main/resources/templates/chatroom.html
@@ -5,6 +5,9 @@
     <title>채팅방 - 북적북적</title>
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Noto+Sans:wght@400;500;700;900&family=Plus+Jakarta+Sans:wght@400;500;700;800&display=swap" />
     <script src="https://cdn.tailwindcss.com?plugins=forms,container-queries"></script>
+
+    <script src="https://cdn.jsdelivr.net/npm/sockjs-client@1/dist/sockjs.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/stompjs@2.3.3/lib/stomp.min.js"></script>
 </head>
 <body class="relative flex min-h-screen flex-col bg-white" style="font-family: 'Plus Jakarta Sans', 'Noto Sans', sans-serif;">
 <div class="flex flex-col grow layout-container">
@@ -12,21 +15,17 @@
 
     <main class="flex-1 flex justify-center px-6 md:px-40 py-6">
         <div class="w-full max-w-[960px] flex flex-col gap-6">
-            <!-- 채팅 상대 및 책 정보 -->
             <div class="border rounded-lg px-4 py-3 bg-gray-50 flex justify-between items-center">
                 <div>
                     <p class="text-lg font-bold" th:text="${partnerNickname}">상대방 닉네임</p>
                     <p class="text-sm text-gray-500" th:text="'책 제목: ' + ${bookTitle}">책 제목</p>
                 </div>
-                <a href="#" class="text-sm text-blue-600 hover:underline">책 보러가기</a>
+                <a th:href="${bookUrl}" class="text-sm text-blue-600 hover:underline">책 보러가기</a>
             </div>
 
-            <!-- 메시지 목록 -->
             <div id="messageArea" class="flex flex-col gap-2 p-4 border rounded-lg h-[400px] overflow-y-auto bg-white shadow-inner">
-                <!-- JS로 메시지들이 실시간 렌더링될 영역 -->
             </div>
 
-            <!-- 입력창 -->
             <div class="flex gap-2">
                 <input id="messageInput" type="text" placeholder="메시지를 입력하세요" class="w-full rounded-lg border border-gray-300 px-4 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-blue-500" />
                 <button id="sendButton" class="px-4 py-2 bg-blue-600 text-white rounded-lg text-sm font-bold hover:bg-blue-700">전송</button>
@@ -38,52 +37,88 @@
 </div>
 
 <script th:inline="javascript">
+    // Thymeleaf에서 받은 변수들을 JS 변수로 초기화
     const channelId = /*[[${channelId}]]*/ 0;
     const currentUserId = /*[[${userId}]]*/ 0;
 
-    const socket = new SockJS("/ws-chat");
+    // SockJS와 STOMP 클라이언트 객체 초기화
+    const socket = new SockJS("/ws-chat"); // WebSocketConfig에 설정된 엔드포인트
     const stompClient = Stomp.over(socket);
 
-    stompClient.connect({}, function () {
-        stompClient.subscribe(`/topic/chat/${channelId}`, function (messageOutput) {
-            const message = JSON.parse(messageOutput.body);
-            showMessage(message);
+    // 에러 콜백 함수 정의
+    const onError = (error) => {
+        console.error('STOMP Error:', error);
+    };
+
+    // 연결 콜백 함수 정의
+    const onConnected = () => {
+        console.log('STOMP Connected!');
+        // 특정 채널을 구독하여 서버로부터 메시지를 받음
+        stompClient.subscribe(`/topic/chat/${channelId}`, (messageOutput) => {
+            showMessage(JSON.parse(messageOutput.body));
         });
 
-        fetch(`/api/chat/channel/${channelId}/messages`)
-            .then(response => response.json())
-            .then(messages => messages.forEach(showMessage));
-    });
+        // 채팅방에 입장하면 이전 대화 기록을 불러옴
+        fetchPreviousMessages();
+    };
 
-    document.getElementById("sendButton").addEventListener("click", () => {
-        const content = document.getElementById("messageInput").value.trim();
-        if (!content) return;
+    // STOMP 클라이언트 연결 실행
+    stompClient.connect({}, onConnected, onError);
 
-        const message = {
-            channelId: channelId,
-            senderId: currentUserId,
-            content: content
-        };
 
-        stompClient.send("/app/chat.send", {}, JSON.stringify(message));
-        document.getElementById("messageInput").value = "";
-    });
+    // 메시지 전송 함수
+    const sendMessage = () => {
+        const messageInput = document.getElementById("messageInput");
+        const content = messageInput.value.trim();
 
-    function showMessage(message) {
-        const msgArea = document.getElementById("messageArea");
+        if (content && stompClient) {
+            const chatMessage = {
+                channelId: channelId,
+                senderId: currentUserId,
+                content: content
+            };
+            // @MessageMapping 경로로 메시지를 전송
+            stompClient.send("/app/chat.send", {}, JSON.stringify(chatMessage));
+            messageInput.value = ""; // 입력창 비우기
+        }
+    };
+
+    // 화면에 메시지를 그리는 함수
+    const showMessage = (message) => {
+        const messageArea = document.getElementById("messageArea");
         const msgEl = document.createElement("div");
 
+        // 내가 보낸 메시지인지 확인
         const isMine = message.senderId === currentUserId;
 
+        // 메시지 스타일 설정
         msgEl.className = `max-w-[70%] px-4 py-2 rounded-lg text-sm shadow ${isMine ? 'bg-blue-100 self-end' : 'bg-gray-100 self-start'}`;
         msgEl.innerText = message.content;
-        msgArea.appendChild(msgEl);
-        msgArea.scrollTop = msgArea.scrollHeight;
-    }
+
+        messageArea.appendChild(msgEl);
+        messageArea.scrollTop = messageArea.scrollHeight; // 항상 최신 메시지가 보이도록 스크롤
+    };
+
+    // 이전 대화 내역을 불러오는 함수
+    const fetchPreviousMessages = () => {
+        fetch(`/api/chat/channel/${channelId}/messages`)
+            .then(response => response.json())
+            .then(messages => {
+                messages.forEach(showMessage);
+            });
+    };
+
+    // '전송' 버튼 클릭 이벤트 리스너
+    document.getElementById("sendButton").addEventListener("click", sendMessage);
+
+    // Enter 키 입력 이벤트 리스너
+    document.getElementById("messageInput").addEventListener("keypress", (e) => {
+        if (e.key === 'Enter') {
+            sendMessage();
+        }
+    });
+
 </script>
 
-<!-- STOMP & SockJS CDN -->
-<script src="https://cdn.jsdelivr.net/npm/sockjs-client@1/dist/sockjs.min.js"></script>
-<script src="https://cdn.jsdelivr.net/npm/stompjs@2.3.3/lib/stomp.min.js"></script>
 </body>
 </html>
